refactor: 패키지 정리 - exception 패키지를 application 안에 넣는다.
