+++ src/main/java/woowacourse/shoppingcart/application/CartService.java
@@ -7,8 +7,8 @@ import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.dao.ProductDao;
 import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.exception.InvalidProductException;
-import woowacourse.shoppingcart.exception.NotInCustomerCartItemException;
+import woowacourse.shoppingcart.application.exception.InvalidProductException;
+import woowacourse.shoppingcart.application.exception.NotInCustomerCartItemException;
 
 import java.util.ArrayList;
 import java.util.List;
