+++ src/main/java/woowacourse/shoppingcart/application/CustomerService.java
@@ -10,8 +10,8 @@ import woowacourse.shoppingcart.dto.CustomerRequest.UserNameAndPassword;
 import woowacourse.shoppingcart.dto.CustomerRequest.UserNameOnly;
 import woowacourse.shoppingcart.dto.CustomerResponse;
 import woowacourse.shoppingcart.dto.DuplicateResponse;
-import woowacourse.shoppingcart.exception.DuplicatedNameException;
-import woowacourse.shoppingcart.exception.InvalidCustomerException;
+import woowacourse.shoppingcart.application.exception.DuplicatedNameException;
+import woowacourse.shoppingcart.application.exception.InvalidCustomerException;
 
 @Service
 @Transactional
