+++ src/main/java/woowacourse/shoppingcart/application/OrderService.java
@@ -7,7 +7,7 @@ import woowacourse.shoppingcart.domain.OrderDetail;
 import woowacourse.shoppingcart.dto.OrderRequest;
 import woowacourse.shoppingcart.domain.Orders;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.exception.InvalidOrderException;
+import woowacourse.shoppingcart.application.exception.InvalidOrderException;
 
 import java.util.ArrayList;
 import java.util.List;
