+++ src/main/java/woowacourse/shoppingcart/application/exception/ShoppingCartException.java
@@ -1,12 +1,12 @@
-package woowacourse.auth.exception;
+package woowacourse.shoppingcart.application.exception;
 
 import org.springframework.http.HttpStatus;
 
-public abstract class AuthException extends RuntimeException {
+public abstract class ShoppingCartException extends RuntimeException {
 
     private final HttpStatus httpStatus;
 
-    public AuthException(String message, HttpStatus httpStatus) {
+    public ShoppingCartException(String message, HttpStatus httpStatus) {
         super(message);
         this.httpStatus = httpStatus;
     }
