+++ src/main/java/woowacourse/shoppingcart/dao/CartItemDao.java
@@ -5,7 +5,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import woowacourse.shoppingcart.exception.InvalidCartItemException;
+import woowacourse.shoppingcart.application.exception.InvalidCartItemException;
 
 import java.sql.PreparedStatement;
 import java.util.List;
