+++ src/main/java/woowacourse/shoppingcart/dao/ProductDao.java
@@ -5,7 +5,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.exception.InvalidProductException;
+import woowacourse.shoppingcart.application.exception.InvalidProductException;
 
 import java.sql.PreparedStatement;
 import java.util.List;
